Remove incorrect fields from AIModDesc
This struct name conflicts with the one in RequiredExports.h.
Additionally, the struts are defined differently. The struct defined
here had an extra two fields.

Checking the assembly code of GetModuleDesc, it's clear those two
addition fields are in error. The struct that is returned is allocated
to be 32 bytes, and so has no room for the two last fields.

This commit is largely to serve as an explanation for why the fields
were removed, and that they were checked and determined to be in error.
Ultimately, this struct will need to go to prevent redefinition errors.