Convert `enum` forward declares to header includes
Forward declaring enums is not technically allowed by the C++ standard.
We would need to have `enum class`, or at least an `enum` with a base
type specific to allow a forward declare.

If client code is expect to use specific `enum` values, then it would be
better to just include the full definition.